Format SQL queries

+++ admin/batch_manager.php
@@ -36,10 +36,7 @@ check_input_parameter('display', $_REQUEST, false, '/^(\d+|all)$/');
 
 if (isset($_GET['action'])) {
     if ($_GET['action'] == 'empty_caddie') {
-        $query = '
-DELETE FROM caddie
-  WHERE user_id = ' . $user['id'] . '
-;';
+        $query = "DELETE FROM caddie WHERE user_id = {$user['id']};";
         pwg_query($query);
 
         $_SESSION['page_infos'] = [
@@ -263,37 +260,23 @@ $filter_sets = [];
 if (isset($_SESSION['bulk_manager_filter']['prefilter'])) {
     switch ($_SESSION['bulk_manager_filter']['prefilter']) {
         case 'caddie':
-            $query = '
-SELECT element_id
-  FROM caddie
-  WHERE user_id = ' . $user['id'] . '
-;';
+            $query = "SELECT element_id FROM caddie WHERE user_id = {$user['id']};";
             $filter_sets[] = query2array($query, null, 'element_id');
 
             break;
 
         case 'favorites':
-            $query = '
-SELECT image_id
-  FROM favorites
-  WHERE user_id = ' . $user['id'] . '
-;';
+            $query = "SELECT image_id FROM favorites WHERE user_id = {$user['id']};";
             $filter_sets[] = query2array($query, null, 'image_id');
 
             break;
 
         case 'last_import':
-            $query = '
-SELECT MAX(date_available) AS date
-  FROM images
-;';
+            $query = 'SELECT MAX(date_available) AS date FROM images;';
             $row = pwg_db_fetch_assoc(pwg_query($query));
             if (! empty($row['date'])) {
-                $query = '
-SELECT id
-  FROM images
-  WHERE date_available BETWEEN ' . pwg_db_get_recent_period_expression(1, $row['date']) . ' AND \'' . $row['date'] . '\'
-;';
+                $date_ = pwg_db_get_recent_period_expression(1, $row['date']);
+                $query = "SELECT id FROM images WHERE date_available BETWEEN {$date_} AND '{$row['date']}';";
                 $filter_sets[] = query2array($query, null, 'id');
             }
 
@@ -301,26 +284,16 @@ SELECT id
 
         case 'no_virtual_album':
             // we are searching elements not linked to any virtual category
-            $query = '
- SELECT id
-   FROM images
- ;';
+            $query = 'SELECT id FROM images;';
             $all_elements = query2array($query, null, 'id');
 
             $linked_to_virtual = [];
 
-            $query = '
- SELECT id
-   FROM categories
-   WHERE dir IS NULL
- ;';
+            $query = 'SELECT id FROM categories WHERE dir IS NULL;';
             $virtual_categories = query2array($query, null, 'id');
             if (! empty($virtual_categories)) {
-                $query = '
- SELECT DISTINCT(image_id)
-   FROM image_category
-   WHERE category_id IN (' . implode(',', $virtual_categories) . ')
- ;';
+                $virtual_categories_ = implode(',', $virtual_categories);
+                $query = "SELECT DISTINCT(image_id) FROM image_category WHERE category_id IN ({$virtual_categories_});";
                 $linked_to_virtual = query2array($query, null, 'image_id');
             }
 
@@ -336,13 +309,7 @@ SELECT id
             break;
 
         case 'no_tag':
-            $query = '
-SELECT
-    id
-  FROM images
-    LEFT JOIN image_tag ON id = image_id
-  WHERE tag_id is null
-;';
+            $query = 'SELECT id FROM images LEFT JOIN image_tag ON id = image_id WHERE tag_id IS NULL;';
             $filter_sets[] = query2array($query, null, 'id');
 
             break;
@@ -372,21 +339,14 @@ SELECT
             // combination of "duplicates_on_fields" you won't get all the
             // duplicates.
 
-            $query = '
-SELECT
-    GROUP_CONCAT(id) AS ids
-  FROM images';
+            $query = 'SELECT GROUP_CONCAT(id) AS ids FROM images';
 
             if (in_array('md5sum', $duplicates_on_fields)) {
-                $query .= '
-  WHERE md5sum IS NOT NULL
-';
+                $query .= ' WHERE md5sum IS NOT NULL';
             }
 
-            $query .= '
-  GROUP BY ' . implode(',', $duplicates_on_fields) . '
-  HAVING COUNT(*) > 1
-;';
+            $duplicates_on_fields_ = implode(',', $duplicates_on_fields);
+            $query .= " GROUP BY {$duplicates_on_fields_} HAVING COUNT(*) > 1;";
             $array_of_ids_string = query2array($query, null, 'ids');
 
             $ids = [];
@@ -402,11 +362,7 @@ SELECT
 
         case 'all_photos':
             if (count($_SESSION['bulk_manager_filter']) == 1) {// make the query only if this is the only filter
-                $query = '
-SELECT id
-  FROM images
-  ' . $conf['order_by'];
-
+                $query = "SELECT id FROM images {$conf['order_by']}";
                 $filter_sets[] = query2array($query, null, 'id');
             }
             break;
@@ -421,11 +377,7 @@ if (isset($_SESSION['bulk_manager_filter']['category'])) {
     $categories = [];
 
     // we need to check the category still exists (it may have been deleted since it was added in the session)
-    $query = '
-SELECT COUNT(*)
-  FROM categories
-  WHERE id = ' . $_SESSION['bulk_manager_filter']['category'] . '
-;';
+    $query = "SELECT COUNT(*) FROM categories WHERE id = {$_SESSION['bulk_manager_filter']['category']};";
     list($counter) = pwg_db_fetch_row(pwg_query($query));
     if ($counter == 0) {
         unset($_SESSION['bulk_manager_filter']);
@@ -438,11 +390,8 @@ SELECT COUNT(*)
         $categories = [$_SESSION['bulk_manager_filter']['category']];
     }
 
-    $query = '
- SELECT DISTINCT(image_id)
-   FROM image_category
-   WHERE category_id IN (' . implode(',', $categories) . ')
- ;';
+    $categories_ = implode(',', $categories);
+    $query = "SELECT DISTINCT(image_id) FROM image_category WHERE category_id IN ({$categories_});";
     $filter_sets[] = query2array($query, null, 'image_id');
 }
 
@@ -452,12 +401,7 @@ if (isset($_SESSION['bulk_manager_filter']['level'])) {
         $operator = '<=';
     }
 
-    $query = '
-SELECT id
-  FROM images
-  WHERE level ' . $operator . ' ' . $_SESSION['bulk_manager_filter']['level'] . '
-  ' . $conf['order_by'];
-
+    $query = "SELECT id FROM images WHERE level {$operator} {$_SESSION['bulk_manager_filter']['level']} {$conf['order_by']}";
     $filter_sets[] = query2array($query, null, 'id');
 }
 
@@ -493,12 +437,8 @@ if (isset($_SESSION['bulk_manager_filter']['dimension'])) {
         $where_clause[] = 'width/height < ' . ($_SESSION['bulk_manager_filter']['dimension']['max_ratio'] + 0.01);
     }
 
-    $query = '
-SELECT id
-  FROM images
-  WHERE ' . implode(' AND ', $where_clause) . '
-  ' . $conf['order_by'];
-
+    $where_clauses_ = implode(' AND ', $where_clause);
+    $query = "SELECT id FROM images WHERE {$where_clauses_} {$conf['order_by']};";
     $filter_sets[] = query2array($query, null, 'id');
 }
 
@@ -513,11 +453,8 @@ if (isset($_SESSION['bulk_manager_filter']['filesize'])) {
         $where_clause[] = 'filesize <= ' . $_SESSION['bulk_manager_filter']['filesize']['max'] * 1024;
     }
 
-    $query = '
-SELECT id
-  FROM images
-  WHERE ' . implode(' AND ', $where_clause) . '
-  ' . $conf['order_by'];
+    $where_clauses_ = implode(' AND ', $where_clause);
+    $query = "SELECT id FROM images WHERE {$where_clauses_} {$conf['order_by']}";
 
     $filter_sets[] = query2array($query, null, 'id');
 }
@@ -586,13 +523,7 @@ $ratios = [];
 $dimensions = [];
 
 // get all width, height and ratios
-$query = '
-SELECT
-  DISTINCT width, height
-  FROM images
-  WHERE width IS NOT NULL
-    AND height IS NOT NULL
-;';
+$query = 'SELECT DISTINCT width, height FROM images WHERE width IS NOT NULL AND height IS NOT NULL;';
 $result = pwg_query($query);
 
 if (pwg_db_num_rows($result)) {
@@ -671,13 +602,7 @@ $template->assign('dimensions', $dimensions);
 $filesizes = [];
 $filesize = [];
 
-$query = '
-SELECT
-  filesize
-  FROM images
-  WHERE filesize IS NOT NULL
-  GROUP BY filesize
-;';
+$query = 'SELECT filesize FROM images WHERE filesize IS NOT NULL GROUP BY filesize;';
 $result = pwg_query($query);
 
 while ($row = pwg_db_fetch_assoc($result)) {
